update by sam

+++ Online-Plant-Nursery/src/main/java/com/cg/online_plant_nursery/dao/FertilizerDAO.java
@@ -9,6 +9,6 @@ import com.cg.online_plant_nursery.entity.Fertilizer;
 @Repository
 public interface FertilizerDAO extends JpaRepository<Fertilizer,Integer>{
 	@Query(value = "from Fertilizer fertilizer where fertilizer.id=?1")
-	 public Fertilizer getFertilizerById(int id);
+	 public Fertilizer getFertilizerById(long id);
 	
 }
