+++ Online-Plant-Nursery/src/main/java/com/cg/online_plant_nursery/dao/PlantDAO.java
@@ -10,6 +10,6 @@ import com.cg.online_plant_nursery.entity.Plant;
 public interface PlantDAO extends JpaRepository<Plant,Integer>{
 
 	@Query(value = "from Plant plant where plant.id=?1")
-	 public Plant getPlantById(int id);
+	 public Plant getPlantById(long id);
 	
 }
